Add to appearance playground
